Recover from bad inputs (#60)
Store and Signal will now catch errors that are thrown when reducers are firing and will prevent the Store from falling apart entirely when errors occur. Store.new now takes an optional error reporter interface, and defaults to printing.

This is a breaking change, as the previous behavior was to propagate the error rather than print it.